refactor: remove keyboard-utils and put everything inside keyboard.js
